+++ src/shared/models/env-var-store/env-vars-info.type.ts
@@ -1,0 +1,7 @@
+import { Version } from "../version.type";
+import { EnvVarsDict } from "../env-var/env-var-data.type";
+
+export type EnvVarStore = {
+  version?: Version;
+  envVars: EnvVarsDict;
+};
