+++ src/shared/models/env-var/env-var-data.type.ts
@@ -1,13 +1,15 @@
+import { Version } from "../version.type";
 import { DefaultValue } from "./default-value.type";
 
 export type EnvVarsDict = { [envVar: string]: EnvVarData };
 
 export type EnvVarData = {
   envVar: string;
-  description: string;
   type: string;
   default?: DefaultValue;
   instances: EnvVarInstance[];
+  description: string;
+  version?: Version;
 };
 
 export type EnvVarInstance = {
