--- src/types/env-vars-info.type.ts
@@ -1,7 +1,0 @@
-import { Version } from "./version.type";
-import { EnvVarsDict } from "../env-var-parser/types/env-var-data.type";
-
-export type EnvVarsInfo = {
-  version?: Version;
-  envVars: EnvVarsDict;
-};
